+++ database/migrations/2020_10_19_160448_create_wards_table.php
@@ -16,7 +16,7 @@ class CreateWardsTable extends Migration
         Schema::create('wards', function (Blueprint $table) {
             $table->id();
             $table->integer('ward_code');
-            $table->integer('district_code');
+            $table->unsignedBigInteger('district_code');
             $table->string('name');
             $table->string('type');
         });
